requirements.txt
astroid==2.1.0
autopep8==1.4.3
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-language==1.1.1
googleapis-common-protos==1.5.6
grpcio==1.18.0
certifi==2018.11.29
chardet==3.0.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nltk==3.4
peewee==3.8.2
protobuf==3.6.1
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pylint==2.2.2
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.7.3
typed-ast==1.3.1
urllib3==1.24.1
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
textblob==0.15.2
urllib3==1.24.1
vaderSentiment==3.2.1
wrapt==1.11.1
xlrd==1.1.0
